+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/schemas/StringSchemaSpec.scala
@@ -253,12 +253,13 @@ class StringSchemaSpec extends UnitSpec {
 
   it should "keep a running histogram of lengths" in {
     val histProp = stringSchema.properties.get[StringLengthHistogramProperty]
-    val bins =
-      (histProp.toJson() \ "lengthHistogram").extract[List[List[Double]]]
+    val histJson = histProp.toJson() \ "lengthHistogram"
+    val bins = (histJson \ "bins").extract[List[List[Double]]]
     bins(0)(0) should equal(4.0 +- 0.1)
     bins(0)(1) should ===(1.0)
     bins(1)(0) should ===(6.0 +- 0.1)
     bins(1)(1) should ===(1.0)
+    (histJson \ "hasExtremeValues").extract[Boolean] shouldBe false
   }
 
   behavior of "StringNumericProperty"