+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Avoid errors with `multipleOf` and very different scales
 - Don't anchor string matches in `StaticPatternProperty`
 - Fix compatibility check with `ProductSchema`
+- Don't crash with extremely large integer values
 
 ## [0.15.0]
 ### Changed